// EcomWeb/src/api/client/models/GetUserInfoResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { GetUserInfoResponse } from './GetUserInfoResponse';
export type GetUserInfoResponseApiResponse = {
    success?: boolean;
    message?: string | null;
    data?: GetUserInfoResponse;
};